Move TabbedPager
