+++ core/src/com/unciv/ui/screens/mapeditorscreen/MapEditorMainTabs.kt
@@ -1,6 +1,6 @@
 package com.unciv.ui.screens.mapeditorscreen
 
-import com.unciv.ui.components.TabbedPager
+import com.unciv.ui.components.widgets.TabbedPager
 import com.unciv.ui.components.input.KeyCharAndCode
 import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.screens.mapeditorscreen.tabs.MapEditorEditTab
