+++ core/src/com/unciv/ui/screens/overviewscreen/UnitOverviewTab.kt
@@ -18,7 +18,7 @@ import com.unciv.models.UpgradeUnitAction
 import com.unciv.models.ruleset.unit.BaseUnit
 import com.unciv.ui.components.widgets.ExpanderTab
 import com.unciv.ui.components.Fonts
-import com.unciv.ui.components.TabbedPager
+import com.unciv.ui.components.widgets.TabbedPager
 import com.unciv.ui.components.UnitGroup
 import com.unciv.ui.components.extensions.addSeparator
 import com.unciv.ui.components.extensions.brighten
