+++ core/src/com/unciv/ui/screens/victoryscreen/VictoryScreenReplay.kt
@@ -6,7 +6,7 @@ import com.badlogic.gdx.utils.Align
 import com.badlogic.gdx.utils.Timer
 import com.unciv.models.UncivSound
 import com.unciv.ui.components.Fonts
-import com.unciv.ui.components.TabbedPager
+import com.unciv.ui.components.widgets.TabbedPager
 import com.unciv.ui.components.UncivSlider
 import com.unciv.ui.components.YearTextUtil
 import com.unciv.ui.components.input.onClick
